+++ lib/model/arg/collection_arguments.dart
@@ -1,0 +1,6 @@
+class CollectionArguments {
+  final int id;
+  final String name;
+
+  CollectionArguments(this.id, this.name);
+}